updated ios sdk to 3.8.5 removed IS_STORE var
