+++ plugin.xml
@@ -17,7 +17,6 @@
     <!-- android -->
     <platform name="android">
         <preference name="ANDROID_API_KEY"/>
-        <preference name="IS_STORE"/>
       
         <config-file target="res/xml/config.xml" parent="/*">
             <feature name="HockeyAppPlugin" >
@@ -37,7 +36,6 @@
         </config-file>
         <config-file target="AndroidManifest.xml" parent="application">
           <meta-data android:name="hockey_app_api_key" android:value="$ANDROID_API_KEY"/>
-          <meta-data android:name="hockey_app_is_store" android:value="$IS_STORE"/>
         </config-file>
 
         <source-file src="src/android/HockeyAppPlugin.java" target-dir="src/org/nypr/cordova/hockeyappplugin/" />
@@ -62,11 +60,9 @@
         <dependency id="org.nypr.cordova.nslogger-cocoalumberjack-connector-plugin" url="https://github.com/wnyc/cordova-plugin-nslogger-cocoalumberjack-connector.git"  />
 
         <header-file src="src/ios/HockeyAppPlugin.h" />
-
         <source-file src="src/ios/HockeyAppPlugin.m" />
         
         <resource-file src="src/ios/HockeySDK.framework/Resources/HockeySDKResources.bundle" />
-        <resource-file src="src/ios/HockeySDK.framework/Resources/HockeySDK.xcconfig" />
 
         <framework src="libc++.tbd" />
         <framework src="src/ios/HockeySDK.framework" custom="true"/>
