--- src/ios/HockeySDK.framework/Resources/HockeySDK.xcconfig
@@ -1,3 +1,0 @@
-OTHER_LDFLAGS=$(inherited) -framework CoreText -framework CoreGraphics -framework Foundation -framework QuartzCore -framework SystemConfiguration -framework UIKit -framework Security -framework AssetsLibrary -framework MobileCoreServices -framework QuickLook -lc++
-HOCKEYSDK_DOCSET_NAME=HockeySDK-iOS
-HOCKEYSDK_GCC_PREPROCESSOR_DEFINITIONS=$(inherited) $(XCODEBUILD_GCC_PREPROCESSOR_DEFINITIONS)
